% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMA_data.R
\name{capture.CMAdata.NEC}
\alias{capture.CMAdata.NEC}
\title{capture CMA meteorological data}
\usage{
capture.CMAdata.NEC(path, YearRange)
}
\arguments{
\item{path}{directory with files downloaded from CMA websites}

\item{YearRange}{Year Range}
}
\value{
a tibble object
}
\description{
capture CMA meteorological data
}
\examples{
path <- "D:/杩呴浄涓嬭浇/姘旇薄鏁版嵁"
YearRange <- c(1957, 2018)
preview.tibble(DS)
}
