% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert.Datetime.excel2R}
\alias{convert.Datetime.excel2R}
\title{convert Excel datetime to R datetime}
\usage{
convert.Datetime.excel2R(x)
}
\arguments{
\item{x}{datetime colum read from Excel file}
}
\description{
convert Excel datetime to R datetime
}
