% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert.pdf2jpg}
\alias{convert.pdf2jpg}
\title{convert pdf to jpeg}
\usage{
convert.pdf2jpg(filefFrom, pages = NULL, dirSaved = NULL, dpi = 300)
}
\arguments{
\item{filefFrom}{pdf file path, a string}

\item{pages}{page number needed to convert}

\item{dirSaved}{directory to save jpeg files}

\item{dpi}{jpeg resolution}
}
\description{
convert pdf to jpeg
}
\examples{
pdf2jpg(pdfpath,dpi = 300)
}
