% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pinyin.R
\name{file.toPinyin}
\alias{file.toPinyin}
\title{Chinese filename to Pinyin filename}
\usage{
file.toPinyin(filename, other_replace = NULL)
}
\arguments{
\item{filename}{a chinese filename}

\item{other_replace}{replace the non chinese character to a specifed char}
}
\value{
a pinyin filename
}
\description{
Chinese filename to Pinyin filename
}
\examples{
citenow()
}
