% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pinyin.R
\name{folder.toPinyin}
\alias{folder.toPinyin}
\title{Convert all filenames in folder to Pinyin files}
\usage{
folder.toPinyin(dir)
}
\arguments{
\item{dir}{a directory containing chinese filenames}
}
\description{
Convert all filenames in folder to Pinyin files
}
\examples{
folder.toPinyin('./')
}
