% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\name{ggplot2.barplot.oneway}
\alias{ggplot2.barplot.oneway}
\title{draw one way bar plot}
\usage{
ggplot2.barplot.oneway(
  trt,
  means,
  std,
  label,
  xlab = NULL,
  ylab = NULL,
  theme.params = NULL
)
}
\arguments{
\item{trt}{the treatment of the one-way experiment}

\item{means}{mean of each treatment}

\item{std}{standard deviation of each treatment}

\item{label}{Significance label of each treatment}

\item{xlab}{x axis title}

\item{ylab}{y axis title}

\item{theme.params}{y theme paramter used to revise ggplot2 theme}
}
\value{
a ggplot2 object
}
\description{
draw one way bar plot
}
\examples{
preview.csv(iris)
}
