% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\name{ggplot2.labs.multiLines}
\alias{ggplot2.labs.multiLines}
\title{add multi-lines xlab and ylab}
\usage{
ggplot2.labs.multiLines(
  g,
  x.labs = NULL,
  y.labs = NULL,
  fontsize = 7,
  padding = 0.4
)
}
\arguments{
\item{g}{a ggplot object}

\item{x.labs}{a vector or list of title for x axis}

\item{y.labs}{a vector or list of title for y axis}

\item{fontsize}{fontsize for both x and y axes}
}
\value{
An arrangeGrob object
}
\description{
add multi-lines xlab and ylab
}
\examples{
g <- ggplot(iris,aes(x=Sepal.Length,y=Sepal.Width,color=Species))+
geom_point()
ggplot2.labs.multiLines(g,x.labs = list("Sepal Length","unit"),
                        fontsize = 7,padding =0.3) \%>\%
Taotao::preview.ggplot(width = 10,height = 7)
 @export
}
