% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{HSDAnalysis.sp.plot}
\alias{HSDAnalysis.sp.plot}
\title{ANOVA for split plot experiment and it's HSD test#'}
\usage{
HSDAnalysis.sp.plot(mplot, splot, block, y)
}
\arguments{
\item{mplot}{vector of the main factor}

\item{splot}{vector of the split factor}

\item{block}{vector of block factor}

\item{y}{vector of a response variable}
}
\description{
ANOVA for split plot experiment and it's HSD test#'
}
\examples{
Packages.InstallandLoad(c('tidyverse','glue'))
}
