% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NECStationINFO}
\alias{NECStationINFO}
\title{Station INFO}
\format{A data frame with 115 obs. and 10 variables:
\describe{
\item{SID} {Station ID in NorthEast China}
\item{Province} {Province Name}
...
}}
\usage{
NECStationINFO
}
\description{
A Data.Frame including Station ID，longitude,latitude, altitude etc.
}
\keyword{datasets}
