% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_and_load.R
\name{Packages.InstallandLoad}
\alias{Packages.InstallandLoad}
\title{install and load packages}
\usage{
Packages.InstallandLoad(pkg)
}
\arguments{
\item{pkgs}{package lists}
}
\description{
install and load packages
}
\examples{
Packages.InstallandLoad(c("tidyverse", "glue"))
}
