% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview.csv}
\alias{preview.csv}
\title{preview dataset using Excel}
\usage{
preview.csv(ds, sheetName = NULL, xl = FALSE, fileEncoding = "gb2312")
}
\arguments{
\item{ds}{a data.frame dataset}

\item{sheetName}{sheet name of excel file}

\item{xl}{whether to use excel.link to preview data.frame}
}
\description{
preview dataset using Excel
}
\examples{
preview.csv(iris)
}
