% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview.ggplot.emf}
\alias{preview.ggplot.emf}
\title{quickly preview ggplot emf}
\usage{
preview.ggplot.emf(Gs, width, height)
}
\arguments{
\item{Gs}{ggplot object}

\item{width}{width of figure (cm)}

\item{height}{height of figure (cm)}
}
\description{
quickly preview ggplot emf
}
\examples{
Gs <- ggplot()+draw_label("abd")
preview.ggplot.emf(Gs,4,4)
}
