% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview.ggplot.pdf}
\alias{preview.ggplot.pdf}
\title{quickly preview ggplot pdf}
\usage{
preview.ggplot.pdf(Gs, width = 10, height = 10)
}
\arguments{
\item{Gs}{ggplot object}

\item{width}{width of figure (cm)}

\item{height}{height of figure (cm)}
}
\description{
quickly preview ggplot pdf
}
\examples{
Gs <- ggplot()+draw_label("abd")
preview.ggplot.pdf(Gs,4,4)
}
