% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview.ggplot}
\alias{preview.ggplot}
\title{quick preview ggplot figure}
\usage{
preview.ggplot(G, width = 17.5, height = 20, dpi = 600)
}
\arguments{
\item{G}{a ggplot2 object}
}
\description{
quick preview ggplot figure
}
\examples{

}
