% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview.latex}
\alias{preview.latex}
\title{quick preview Latex expression}
\usage{
preview.latex(tex.strings = "$y=x^2+x \\\\times y$")
}
\arguments{
\item{tex.strings}{a laTex list}
}
\description{
quick preview Latex expression
}
\examples{
latex.preview(list("$y=x^2+1,r^2=0.96,p<0.05$", "$\\\\overset{y=x^2+1}{r^2=0.96,p<0.05}{a}$"))
latex.preview("$\\\\overset{grain\\\\,yield\\\\,grain\\\\,yield}{(kg\\\\,ha^{-1})}$")
 @export

}
