% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview.MathExpressions}
\alias{preview.MathExpressions}
\title{quickly preview Mathematical Expression}
\usage{
preview.MathExpressions(expr1, ...)
}
\arguments{
\item{exp}{expression}
}
\description{
quickly preview Mathematical Expression
}
\examples{
expr1 = expression(c^2 == a[1]^2 + b[1]^2)
preview.MathExpressions(expr1)
}
