% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview.tibble}
\alias{preview.tibble}
\title{data visualization using shiny}
\usage{
preview.tibble(data_tibble)
}
\arguments{
\item{tibble}{a data.frame dataset}
}
\description{
data visualization using shiny
}
\examples{
tibble_preview(iris)
}
