% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview.tmap}
\alias{preview.tmap}
\title{quick preview tmap figure}
\usage{
preview.tmap(..., ncol = 1, nrow = 1, width = 10, height = 7, res = 600)
}
\description{
quick preview tmap figure
}
\examples{

}
