% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview.word_table}
\alias{preview.word_table}
\title{preview Word Table}
\usage{
preview.word_table(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
preview Word Table
}
\examples{

preview.word_table(iris)
 @export

}
