% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{HSDAnalysis.sp.plot}
\alias{HSDAnalysis.sp.plot}
\title{ANOVA for split plot experiment and it's HSD test#'}
\usage{
HSDAnalysis.sp.plot(DS, block, pplot, splot)
}
\arguments{
\item{DS}{data.frame dataset}

\item{block}{a string block factor Name}

\item{pplot}{a string main factor factor Name}

\item{splot}{a string split factor factor Name}
}
\description{
ANOVA for split plot experiment and it's HSD test#'
}
\examples{
Packages.InstallandLoad(c('tidyverse','glue'))
}
